+++ apps/web/lib/api/invoices/utils.ts
@@ -1,5 +1,5 @@
 import { prisma } from "@dub/prisma";
-import { randomBytes } from "crypto";
+import { generateRandomString } from "@dub/utils";
 
 export async function generateInvoicePrefix() {
   const invoicePrefix = generateRandomString(8);
@@ -15,19 +15,4 @@ export async function generateInvoicePrefix() {
   }
 
   return invoicePrefix;
-}
-
-// TODO:
-// Move this to shared utils
-function generateRandomString(length: number): string {
-  const charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-  const randomBytesArray = randomBytes(length);
-  let result = "";
-
-  for (let i = 0; i < length; i++) {
-    const randomIndex = randomBytesArray[i] % charset.length;
-    result += charset[randomIndex];
-  }
-
-  return result;
 }
